# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /user/createUser:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: rePassword
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户
  /user/deleteUser:
    post:
      parameters:
      - description: 用户Id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户
  /user/getUserByNameAndPassword:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 根据用户名和密码查询用户
      tags:
      - 用户
  /user/getUserList:
    get:
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 用户列表
      tags:
      - 用户
  /user/updateUser:
    post:
      parameters:
      - description: 用户Id
        in: formData
        name: id
        type: string
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      - description: 电话
        in: formData
        name: phone
        type: string
      - description: 邮箱
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code", "message"}
          schema:
            type: string
      summary: 更新用户
      tags:
      - 用户
swagger: "2.0"